requirements_all.txt
# Requirements for updated_sales.py
squareup>=29.0.0.20230720
python-dateutil>=2.8.0
pytz>=2021.1
